feat: adicion chamada de cardProducts ao arquivo apiGets
